# Optimization of LTC signals