--- a/core/federated/federate.c
+++ b/core/federated/federate.c
@@ -2201,9 +2201,9 @@ void* lf_handle_p2p_connections_from_federates(void* env_arg) {
 
 void lf_latest_tag_confirmed(tag_t tag_to_send) {
   environment_t* env;
-  int compare_with_last_LTC = lf_tag_compare(_fed.last_sent_LTC, tag_to_send);
-  if (compare_with_last_LTC >= 0) {
+  if (lf_tag_compare(_fed.last_sent_LTC, tag_to_send) >= 0) {
     return; // Already sent this or later tag.
+  }
   }
   _lf_get_environments(&env);
   if (!env->need_to_send_LTC) {