--- a/core/threaded/reactor_threaded.c
+++ b/core/threaded/reactor_threaded.c
@@ -880,6 +880,15 @@ void _lf_worker_do_work(environment_t* env, int worker_number) {
                    worker_number, current_reaction_to_execute->name, LF_LEVEL(current_reaction_to_execute->index),
                    current_reaction_to_execute->is_an_input_reaction, current_reaction_to_execute->deadline);
 
+    #ifdef FEDERATED_CENTRALIZED
+      if (current_reaction_to_execute->is_an_input_reaction) {
+        // This federate has received a tagged message with the current tag and 
+        // must send LTC at the current tag to confirm that the federate has successfully
+        // received and processed tagged messages with the current tag.
+        env->need_to_send_LTC = true;
+      }
+    #endif // FEDERATED_CENTRALIZED
+
     bool violation = _lf_worker_handle_violations(env, worker_number, current_reaction_to_execute);
 
     if (!violation) {