--- a/include/core/environment.h
+++ b/include/core/environment.h
@@ -87,10 +87,8 @@ typedef struct environment_t {
 #if defined(FEDERATED)
   tag_t** _lf_intended_tag_fields;
   int _lf_intended_tag_fields_size;
-#endif             // FEDERATED
-#if defined(FEDERATED_CENTRALIZED)
   bool need_to_send_LTC;
-#endif             // FEDERATED_CENTRALIZED
+#endif // FEDERATED
 #ifdef LF_ENCLAVES // TODO: Consider dropping #ifdef
   enclave_info_t* enclave_info;
 #endif