We have a companion PR in [lf-lang/lingua-franca#2346](https://github.com/lf-lang/lingua-franca/pull/2346).

Currently, a federate sends the latest tag completed (LTC) signals every time the federate completes a tag. However, this is not necessary as the primary goal of LTC signals is to confirm that the federate has received and processed tagged messages with the current tag (and tags earlier than the current tag). Thus, I suggest renaming the signal's name to the latest tag confirmed (the abbreviation is still LTC) and making federates send the LTC signal only if the federates have received any tagged message with the current tag. 

The other role of LTC signals is to allow its "immediate" downstream federates to advance their tags. The only downside of this change is that the RTI may take more time to compute TAG signals. For example, let's consider an example as described below.

![image](https://github.com/lf-lang/reactor-c/assets/78055940/6ff6b3c4-149a-4f7c-8f83-b313797050a5)

In the current implementation, federate B sends `LTC (100 msec)` when B completes `100 msec`. Then, the RTI can grant `TAG (100 msec)` with `LTC (100 msec)`. However, if B does not send `LTC (100 msec)` (because A does not send any tagged messages with `100 msec`), the RTI has to look up A as well as B to calculate C's `EIMT` to grant `TAG (100 msec)`. Yet, I argue that this impact is negligible because we almost never see a million federates connected together. 

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Refactor**
  - Renamed functions and updated log messages from "latest tag complete" (LTC) to "latest tag confirmed" (LTC) to clarify event meaning.
  
- **New Features**
  - Introduced a conditional logic to handle federated centralized mode, setting a flag for message processing confirmation.

- **Chores**
  - Updated comments and variable names for better clarity and consistency in the code.
  - Changed branch name reference from "master" to "LTC-optimization."
<!-- end of auto-generated comment: release notes by coderabbit.ai -->