@lhstrh 2024-07-01
I noticed a segfault in the tests. Let's defer merging this until that has been sorted out...