@byeonggiljun 2024-07-31
@lhstrh, I resolved every conversation except for the one regarding `lingua-franca-ref.txt` which should be resolved after the PR [lingua-franca#2346](https://github.com/lf-lang/lingua-franca/pull/2346) being merged.